pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aos"
version = "1.0.0"
description = "AOS — Agentic Operating System. Sovereign LLM Gateway with energy-aware model routing."
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
rag = [
    "llama-index>=0.12.0",
    "llama-index-vector-stores-postgres>=0.4.0",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-llms-ollama>=0.5.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
aos = "aos.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
